Revert "Make sure the binding is only initialized once"
Sorry the reverted commit did not fixed the problem of init registering
subscriptions on selection, anchor and focused therefore I'll revert it.
The commit will not do any harm, but I will not help eighter. We need to
wrap these observables or upgrade to Knockout 2.2 and use peek. I'll sto
p experimenting on the master branch now - sorry.

This reverts commit fc92e1968dbda801c5d8c1fc788cb82d5eff8792.